// apps/foreldrepengeoversikt/src/app/types/Skjemanummer.ts
export enum Skjemanummer {
    ANNET = 'I000060',
    BEKREFTELSE_DELTAR_KVALIFISERINGSPROGRAM = 'I000051',
    BEKREFTELSE_FRA_ARBEIDSGIVER = 'I000065',
    BEKREFTELSE_FRA_STUDIESTED = 'I000061',
    BEKREFTELSE_PÅ_AVTALT_FERIE = 'I000036',
    DOK_AV_ALENEOMSORG = 'I000110',
    DOK_BEGRUNNELSE_SØKE_TILBAKE_I_TID = 'I000111',
    DOK_DELTAKELSE_I_INTRODUKSJONSPROGRAMMET = 'I000112',
    DOK_INNLEGGELSE = 'I000037',
    DOK_MILITÆR_SILVIL_TJENESTE = 'I000039',
    DOK_MORS_UTDANNING_ARBEID_SYKDOM = 'I000038',
    DOK_OVERFØRING_FOR_SYK = 'I000023',
    ETTERLØNN_ELLER_SLUTTVEDERLAG = 'I000044',
    FØDSELSATTEST = 'I000063',
    INNTEKTSOPPLYSNINGER_FRILANS_ELLER_SELVSTENDIG = 'I000007',
    OMSORGSOVERTAKELSESDATO = 'I000042',
    SKJEMA_FOR_TILRETTELEGGING_OG_OMPLASSERING = 'I000109',
    TERMINBEKREFTELSE = 'I000062',
    TILBAKEBETALING = 'I000114',
    TILRETTELEGGING_FOR_ARBEIDSTAKERE = 'I000049',
    TILRETTELEGGING_FOR_FRILANS_ELLER_SELVSTENDIG = 'I000001',
    HV_ØVELSE = 'I000116',
    NAV_TILTAK = 'I000117',
    RESULTATREGNSKAP = 'I000032',
    KOPI_SKATTEMELDING = 'I000066',
}